requirements.txt
# Core
aiogram==3.22.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Caching (опціонально)
redis==5.0.1

# Utilities
pydantic==2.5.2
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-env==0.8.2

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
